chore: update time slot structure
