+++ src/screens/index.ts
@@ -3,8 +3,8 @@ export { default as SearchScreen } from './Tabs/SearchTab/Search/SearchScreen';
 export { default as FavouritesScreen } from './Tabs/FavouritesTab/Favourites/FavouritesScreen';
 export { default as CartScreen } from './Tabs/CartTab/Cart/CartScreen';
 export { default as ProfileScreen } from './Tabs/ProfileTab/Profile/ProfileScreen';
-export { default as TimeSlotPickerModalScreen } from './Schedule/TimeSlotPicker/TimeSlotPickerModalScreen';
-export { default as TimeSlotPickerExampleScreen } from './Schedule/TimeSlotPicker/TimeSlotPickerExampleScreen';
-export { default as TimeSlotPickerComponentOnlyScreen } from './Schedule/TimeSlotPicker/TimeSlotPickerComponentOnlyScreen';
+export { default as TimeSlotPickerModalScreen } from '~screens/Schedule/TimeSlotPickerScreens/TimeSlotPickerModalScreen';
+export { default as TimeSlotPickerExampleScreen } from '~screens/Schedule/TimeSlotPickerScreens/TimeSlotPickerExampleScreen';
+export { default as TimeSlotPickerComponentOnlyScreen } from '~screens/Schedule/TimeSlotPickerScreens/TimeSlotPickerComponentOnlyScreen';
 export { default as EntityScreen } from './Entity/EntityScreen';
 export { default as CheckoutScreen } from './Checkout/CheckoutScreen';
